Create new qcow2ssh tm
The ssh tm allows host-local storage to be used for vm's. The qcow2 tm
ensures that all images are qcow2 so can use those features such as
quick snapshotting. This tm combines the two with an improved snapshot
deletion procedure which allows a tree of snapshots to be managed as if
each point is a full clone of the image at that point in time.